Events List Domain + Unit Test

+++ lib/features/events/domain/entities/event.dart
@@ -1,8 +1,9 @@
 import 'dart:ffi';
 
 import 'package:freezed_annotation/freezed_annotation.dart';
-import 'package:rg_rek/features/events/domain/entities/event_extended_data.dart';
-import 'package:rg_rek/features/events/domain/entities/event_game.dart';
+
+import 'event_extended_data.dart';
+import 'event_game.dart';
 
 part 'event.freezed.dart';
 
@@ -23,7 +24,7 @@ class Event with _$Event {
     required DateTime eventStart,
     required int eventType,
     required int gamesCount,
-    required Long remoteId,
+    required int remoteId,
     required List<EventExtendedData> eventExtendedData,
     required List<EventGame> eventGames,
   }) = _Event;
